Lab_1/Titanic/requirements.txt
hopsworks
joblib
scikit-learn
seaborn
dataframe-image
modal-client
gradio
mimesis